+++ examples/replay.hs
@@ -1,39 +1,18 @@
-{- |
-Description : Replay configuration example.
-
-This module is an example replay configuration.
-
-Usage:
-
-To run a program:
-vcr custom --program "</abs/path/to/this/script> --asterix <abs/path/to/asterix/files>" --run
-
-To check:
-vcr custom --program "</abs/path/to/this/script>" --validate
--}
+-- | Description : Replay configuration example.
 
 {-# LANGUAGE OverloadedStrings #-}
 
 -- standard imports
-import           Data.Bool
-import qualified Data.ByteString  as BS
-import           Data.Char        (toUpper)
-import           Data.List        (nub)
-import           Data.Maybe
-import qualified Data.Text        as T
+import qualified Data.ByteString as BS
+import           Data.Char       (toUpper)
+import qualified Data.Text       as T
 import           Data.Word
-import qualified Pipes.Prelude    as PP
-import qualified Pipes.Safe       as PS
-import           System.Directory (getDirectoryContents)
-import           System.FilePath  ((</>))
+import qualified Pipes.Prelude   as PP
+import qualified Pipes.Safe      as PS
 import           Text.Printf
 
 -- VCR imports
 import           Replay
-
--- Asterix processor
-import qualified Data.Asterix     as Ast
-import qualified Data.BitString   as Bits
 
 -- List of recorders.
 recorders :: [(Name, Source)]
@@ -80,57 +59,6 @@ dump evt = printf "%s: %-10s: 0x%-10s...\n"
 prepend :: MonadIO m => Word8 -> Pipe BS.ByteString BS.ByteString m ()
 prepend b = PP.map (\bs -> BS.singleton b <> bs)
 
--- | Replay filter example.
--- It manipulates timestamp in some asterix records,
--- such that it looks like the record was just generated.
--- Take into consideration original delay when restamping.
-dgRestamp :: MonadIO m => Ast.Profiles -> Pipe (Event UdpContent) BS.ByteString m ()
-dgRestamp uaps = forever $ do
-    event <- await
-    now <- tod <$> liftIO getUtcTime
-    yield $ processDatagram
-        now
-        (tod $ eTimeUtc event)
-        (udpDatagram $ eValue event)
-  where
-
-    tod :: UtcTime -> Ast.EValue
-    tod t = Ast.EDouble $ fromRational $ toRational $ utctDayTime t
-
-    processDatagram :: Ast.EValue -> Ast.EValue -> BS.ByteString -> BS.ByteString
-    processDatagram now oldUtc s = case Ast.toDataBlocks $ Bits.fromByteString s of
-        Nothing -> s    -- not an asterix
-        Just dbs ->     -- process each datablock, encode, check alignment
-            let result = mconcat $ fmap (Ast.fromDataBlock . processDataBlock) dbs
-            in bool s (Bits.toByteString result) (Bits.isAligned result)
-      where
-        processDataBlock :: Ast.DataBlock -> Ast.DataBlock
-        processDataBlock db = case mProcessRecord of
-            Nothing -> db
-            Just processRecord -> case Ast.toRecords uaps db of
-                Nothing -> db
-                Just recs -> Ast.mkDataBlock (Ast.dbCat db) (fmap processRecord recs)
-          where
-            mProcessRecord :: Maybe (Ast.Item -> Ast.Item)
-            mProcessRecord = fmap modifyItem itemToModify
-              where
-                itemToModify = case Ast.dbCat db of
-                    1  -> Just "141"
-                    2  -> Just "030"
-                    19 -> Just "140"
-                    20 -> Just "140"
-                    34 -> Just "030"
-                    48 -> Just "140"
-                    _  -> Nothing
-                modifyItem i rec = fromMaybe rec $ Ast.update rec $ do
-                    Ast.modifyItem i $ \j dsc -> do
-                        t1 <- Ast.toRaw j
-                        t2 <- Ast.fromNatural oldUtc dsc >>= Ast.toRaw
-                        t3 <- Ast.fromNatural now dsc >>= Ast.toRaw
-                        let originalDelay = t2 - t1
-                            t4 = t3 - originalDelay
-                        Ast.fromRaw (t4 :: Int) dsc
-
 -- | Extract datagram from event.
 dg :: Monad m => Pipe (Event UdpContent) BS.ByteString m ()
 dg = PP.map (udpDatagram . eValue)
@@ -158,16 +86,14 @@ noBlink :: Event UdpContent -> Maybe Double
 noBlink _event = Nothing
 
 -- | Replay sessions.
-outputs ::
-    Ast.Profiles        -- asterix profiles
-    -> [ (Name      -- session name
-       , [ ( Channel        -- channel name
-           , BlinkTime      -- seconds to switch on active indicator
-           , ConsoleDump    -- console dump function
-           , Output         -- event consumer
-           )]
-       )]
-outputs uaps =
+outputs :: [ (Name      -- session name
+    , [ ( Channel        -- channel name
+        , BlinkTime      -- seconds to switch on active indicator
+        , ConsoleDump    -- console dump function
+        , Output         -- event consumer
+        )]
+    )]
+outputs =
     [ ("normal",   -- normal replay
         [ ("ch1", blink 1.0, dump, txUnicast dg "127.0.0.1" "59001")
         , ("ch2", blink 0.1, dump, txUnicast dg "127.0.0.1" "59002")
@@ -181,12 +107,6 @@ outputs uaps =
         , ("ch3", blink 1.0, dump, txUnicast (dg >-> prepend 3) "127.0.0.1" "59003")
         , ("ch4", blink 1.0, dump, txUnicast (dg >-> prepend 4) "127.0.0.1" "59004")
         ])
-    , ("restamp",   -- restamp asterix
-        [ ("ch1", blink 1.0, dump, txUnicast (dgRestamp uaps) "127.0.0.1" "59001")
-        , ("ch2", blink 1.0, dump, txUnicast (dgRestamp uaps) "127.0.0.1" "59002")
-        , ("ch3", blink 1.0, dump, txUnicast (dgRestamp uaps) "127.0.0.1" "59003")
-        , ("ch4", blink 1.0, dump, txUnicast (dgRestamp uaps) "127.0.0.1" "59004")
-        ])
     , ("multicast",   -- send to multicast
         [ ("ch1", blink 1.0, dump, txMulticast dg "239.0.0.1" "59001" "127.0.0.1" 32)
         , ("ch2", blink 1.0, dump, txMulticast dg "239.0.0.1" "59002" "127.0.0.1" 32)
@@ -195,47 +115,9 @@ outputs uaps =
         ])
     ]
 
-newtype Options = Options
-    { optAstData    :: [FilePath]
-    } deriving (Show)
-
-options :: Parser Options
-options = Options
-    <$> some (strOption (long "asterix" <> help "Asterix data definition directory"))
-
 -- A main program, start replay GUI.
 main :: IO ()
 main = do
-    opt <- execParser (info (options <**> helper) idm)
-
-    -- read asterix definition files
-    uaps <- do
-        let paths = optAstData opt
-        files <- collectFiles [".xml", ".XML"] paths
-        when (null files) $
-            error $ "no xml files found in " ++ show paths
-        descriptions <- forM files $ \filename -> do
-            xml <- readFile filename
-            case Ast.categoryDescription xml of
-                Left msg  -> error $ filename ++ ", " ++ msg
-                Right val -> pure val
-        -- get latest revisions
-        case Ast.categorySelect descriptions [] of
-                Left msg  -> error msg
-                Right val -> pure val
-
     runReplay
         (50*1000) -- console buffer size
-        recorders channelMaps (outputs uaps)
-
--- | Get specific files from list of directories.
-collectFiles :: [String] -> [FilePath] -> IO [FilePath]
-collectFiles extensions dirs = nub . filter match . concat <$> mapM ls dirs
-  where
-    ls dir = do
-        listing <- filter (`notElem` [".", ".."]) <$> getDirectoryContents dir
-        pure $ map (dir </>) listing
-    match filename = any (matchSuffix filename) extensions
-    matchSuffix filename suffix
-        | length filename < length suffix = False
-        | otherwise = and $ zipWith (==) (reverse filename) (reverse suffix)
+        recorders channelMaps outputs