+++ CHANGELOG.md
@@ -2,4 +2,10 @@
 
 ## 0.1.0 -- YYYY-mm-dd
 
-* First version. Released on an unsuspecting world.
+* First version.
+
+## 2.0.0 -- 2024-10-03
+
+* Nix packages update.
+* Project structure cleanup, lint and style update.
+* 'custom' command removed (replay script moved to a separate vcr-config).